// frontend/src/app/api/model/grupo.ts
/**
 * API oPROsita
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Mes } from './mes';


export interface Grupo { 
    id?: number;
    nombre?: string;
    meses?: Array<Mes>;
}