// frontend/src/app/api/model/novedad.ts
/**
 * API oPROsita
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Novedad { 
    id?: number;
    texto?: string;
    tipoDestinatario?: Novedad.TipoDestinatarioEnum;
    fechaCreacion?: string;
}
export namespace Novedad {
    export const TipoDestinatarioEnum = {
        Profesor: 'profesor',
        Alumno: 'alumno'
    } as const;
    export type TipoDestinatarioEnum = typeof TipoDestinatarioEnum[keyof typeof TipoDestinatarioEnum];
}